+++ frontend/src/Context/ConfigProvider/useConfig/index.tsx
@@ -1,0 +1,4 @@
+import useConfig, { defaultConfig } from './useConfig';
+
+export { defaultConfig };
+export default useConfig;
