--- frontend/src/Context/ConfigProvider/useConfigProvider/index.tsx
@@ -1,4 +1,0 @@
-import useConfig, { defaultConfig } from './useConfigProvider';
-
-export { defaultConfig };
-export default useConfig;
